test: 모임 POJO 테스트 추가 (#191)
